requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.2.4
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==23.1
phonenumbers==8.13.18
pluggy==1.2.0
pytest==7.4.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
